Classifier tasks: Add SubjectViewerStore to task stories (#2031)
Rewrite all task stories as CSF. Move common task story code into a `MockTask` component, which takes a `tasks` object and renders it in a `Tasks` component, with a classifier store. Adds `@stories` as an alias for storybook helper code and `@components` as an alias for the root of the Classify components tree. Fixes a small bug where feature detection helpers weren't being exported properly.