Classifier: add support for recursive, branching workflows (#2050)
* Bump mst-middlewares to 3.17

* Add annotation history to workflow steps
Add AnnotatedSteps, which records a history of completed workflow steps, along with their annotations. The Next button records the latest completed step, along with its annotations. The Back button appears if undo is possible, and undoes the last step. getPreviousStepKey and isThereAPreviousStep are removed, because they aren't used anywhere.

* Support conditional branching and recursion
Add the history manager to the Tasks component. Task annotations are rendered from the history stack if this is a step that we’ve previously completed.

The Back button undoes the current step. The Next button adds a new step to the stack if we are starting a step for the first time. 

When the subject queue advances, the stack is cleared and a new stack started from the first workflow step. The first step in the stack cannot be undone.

Remove unused branching code from the WorkflowSteps and Step models.

* Remove inject from Tasks components
Add MobXContextProvider to Tasks, Task and TaskNavButtons.

* Generalise task.isComplete
Change task.isComplete to task.isComplete(annotation), so that multiple annotations are supported for the same task.
Remove cruft from the drawing task and transcription task tests.

* Clean up task.setAnnotation
Remove task.setAnnotation(annotation) from models, examples and tests. Mark task.annotation as deprecated in the Readme.

* Change responsibility for selecting steps
Explicitly change the active step by calling workflowSteps.selectStep() from consumers. Remove implicit step changes from the workflow step actions.

Co-authored-by: Shaun A. Noordin <shaunanoordin@users.noreply.github.com>