Classifier drawing tools: fix focus and window scroll position for new marks (#2080)
* Remove onFinishWithRef
Remove `onFinishWithRef` from the `DrawingToolMarks` component. Set subtask visibility directly when a new mark is finished, or when an existing mark is selected. Remove `setSubTaskVisibility` from the `onFinish` handler.

* Update Mark tests
Update Mark tests to use mark models, not snapshots, and to make sure that the subtask cases use a tool with subtasks.

* Remove releasePointerCapture
The browser handles pointer release for us, on pointer up. Our code may throw invalid pointer errors if it tries to release the pointer itself.

* Add openSubTaskPopup to marks
Add openSubTaskPopup, which checks for finished marks, with subtasks, before opening the popup. Call it after creating a new mark, on Enter or Space and on pointerup for existing marks.

* Pass mark bounds, not a ref
Pass the current marks bounding client rectangle when setting subtask visibility.

* Focus marks with subtasks

* Focus transcribed lines on creation
Split out setting previous mark annotations and showing the subtask popup into two separate actions. Finish new transcription lines after adding previous transcriptions to the line. Check text tasks for an existing value before showing the suggestions dropdown.

* Fix scroll position on popup open
Remember the window scroll position when the subtask popup opens, and fix the window scroll to that position.

* Observe marks for subtask visibility changes
If subtask visibility changes from true to false, focus the mark and restore the window scroll position.

* Scroll selected marks into view

* Allow marks to use keyboard zoom
Add `g` to the list of allowed tags in withKeyZoom.

Co-authored-by: srallen <srallen@users.noreply.github.com>