Refactor task navigation (#2062)
* Refactor BackButton
Split BackButton into a view and a HOC that connects it to the store. Copy over behaviour tests for the HOC from TaskNavButtonsContainer.

* Refactor DoneAndTalkButton
Split DoneAndTalkButton into a view and a HOC that connects it to the store. Copy over behaviour tests for the HOC from TaskNavButtonsContainer.

* Refactor DoneButton
Split DoneButton into a view and a HOC that connects it to the store. Copy over behaviour tests for the HOC from TaskNavButtonsContainer.

* Refactor NextButton
Split NextButton into a view and a HOC that connects it to the store. Copy over behaviour tests for the HOC from TaskNavButtonsContainer.

* Remove TaskNavButtonsContainer
Render buttons directly from the TaskNavButtons component.

* Test Done buttons when there's a next step

* Add step.completeTasks(annotations)

* Simplify NextButtonConnector
Move step.completeTasks(annotations) into the next() action, which simplifies the store connector.

* Add annotatedSteps.finish()
Add a finish action, complementing start(), which runs at the end of a classification, when pressing Done or Done & Talk.

* Test step.completeTasks
Test that completing a step also completes that step's tasks.

Co-authored-by: srallen <srallen@users.noreply.github.com>