Engaging Crowds: Subject picker tool (#1925)
* Add the subject picker to workflow menus
Add the SubjectPicker component to the workflow menus for both the project home page and the project classify page. RedSea only has one indexed subject set at the moment, so I've hardcoded the picker to workflow 16106 and subject set 15582. When you select a subject set for workflow 16106, subjects are requested from RedSea and shown in a Grommet DataTable. The data table can be used to sort and filter subjects, before choosing one to classify.