Project: remove inject from the classifier wrapper (#2176)
* Project: remove inject from the classifier wrapper
Refactor the Classifier wrapper to use MobXProviderContext instead of inject. I've used a HOC pattern to add store props, so that the tests can continue to run against the undecorated component.

* Split out ClassifierWrapper from the container
Split the container into two components. Add some initial documentation, with an example.
Test the container, with MobX context
Rename container to connector
Add a useStore hook
Refactor the check for a logged-in user
Check for a logged-in user in the parent component, so that we re-render when auth completes.