Rotating rectangle tool (#2211)
* ESlint/Prettier fixes per Zooniverse Style Guide

* Created RotateRectangle from Rectangle

* WIP RotateRectangle Marks model

* Debugging RotateRectangle marks model

* Removed inherited code

* WIP save

* Added the rotate guide line

* Created the rotateHandle

* Created onRotateDrag function

Fixed rectangle onHandleDrag bug: added the angle

* Change RotateRectangle to use Mark's transform=translate to position itself and move the origin point to (x,y)

* Created x_roate and y_rotate

In the transform, we now check for a x_rotate and y_rotate to use as a pivot point for rotating.

* Created unit test

* Code clean-up

Declared the rotate() transform string as a variable to make the code easier to read.

* Removed ‘only’

Was being used during developing.

* Added observer for RotateRectangle

DrawingToolMarks no longer wraps each drawing tool. We must explicitly wrap each drawing tool with an observer to properly render and re-render the svg when resizing, reshaping.

* Created new RotateHandle component

to use for tools having a rotate feature. The handle is a SVG element.

* Updated unit test for RotateHandle

* Added unit test for draging the RotateHandle

* Refactored to use ‘beforeEach’

* Added unit test for onFinish

* Updated RotateHandle for Storybook documentation

* Removed _onFinish

In favor of default onFinish which is passed down.

* Updated RotateRectangle

for Storybook documentation

* Removed children

‘children’ is no longer used. Was left over from an earlier version of the drawing tools, where the delete button was inserted into each tool as a child.

* Defined ‘mark’ with PropTypes.shape

to type check and better define what was in ‘mark’.
Alphabetized properties of ‘mark’ cause of my OCD! :)

Co-authored-by: Shaun A. Noordin <shaun.a.noordin@gmail.com>