Engaging Crowds: Add URL workflow and subject set to WorkflowMenu (#2169)
* Add URL workflow and subject set to WorkflowMenu
If workflowID and/or subjectSetID are defined in the Classify page URL, this skips past workflow selection and/or subject set selection when opening the workflow menu.

* Test the workflow menu logic
Test the workflow menu logic with different combinations of props. Add some explanatory comments.

* Defer setting classify props until ready
Wait until canClassify is true before setting workflowID, subjectSetID or subjectID in the classifier. Otherwise, subject queries can be sent with incomplete URLs.