Use UTC for daily stats calculations (#2359)
* Use UTC for daily stats calculations
Ignore the local time zone and use UTC when calculating your daily classification counts. With this change, the daily stats chart should always begin on a Monday, regardless of which time zone you're in. Classifications should be shown on the day that you made them, not the day before.

* Convert stories to CSF

* Remove inject

* Update story component name

* Add the timeZone option
Use UTC when getting day names for dates from the stats server.

* Fix the check for the current day number
Today's day number should be calculated from the local clock. Day numbers for stats data should use UTC.

* Refactor the daily counts
Calculation of the daily counts is a bit hard to follow. This might help clarify it.

* Fix test expectations
Don't localise the expected day names. They should be the same, regardless of which time zone and language the tests are running in.
Run the mock clock at noon.