Engaging Crowds: Next/Previous subject navigation (#2392)
* Add desc sort to getIndexedSubjects
Allow getIndexedSubjects to return previous subjects, ordered in descending priority. Return the subjects array for greater flexibility.

* New subjectStore actions and properties:
- `subjects.queue`: an array of Subject IDs, in the order that they'll be classified. New subjects can be appended to the end of the queue, or prepended to the start of the queue.
- `nextIndexed`: select the next subject, by priority.
- `previousIndexed`: select the previous subject, by priority.
- `setActiveSubject(subjectID)`: set the active subject, by ID.
- `shift`: shift the queue by one subject and make the first subject active.
- `splice(newSubjects)`: interpolate `newSubjects` into the queue, in priority order.

* Update the store on `setActiveSubject`

* Add subjects.buildPreviousQueue
Build the previous queue asynchronously, so that we can update the active subject immediately, if possible.

* Add navigation buttons
Add navigation buttons to the subject progress banner. Wire them up to `subjects.previous` and `subjects.next`.

* Add subjects.queue
subjects.queue is an array of subject IDs, in the order that we want to show them to a volunteer.

* Remove `subjects.clearAvailable`

* Skip all the failing Choice tests