Fix broken drawing tool stories (#2498)
* Fix broken drawing tool stories
Mock a full store for the drawing tool stories, so that they don't break when new features are added to the store but are missing from the mock store in these stories.

* Fix TranscribedLines
Add the mock store to TranscribedLines stories. Allow the mock API client to be configured by the calling code.

* Convert TranscribedLines to CSF

* Remove withKnobs

* Fix mock mark bounds

* Use a larger mock subject

* Setup shared helpers
Move duplicate story setup code into `drawingTools/stories/helpers`.

* Fix draggable behaviour
Set the viewer loading state to success, so that `canvas` is defined.

* Allow mark to be undefined
`mark` is undefined if you delete all the example marks from a story.

* Use a valid subject
Subject URLs must end in a valid extension, so use a Panoptes subject URL and add the mock subject to the store.

* Wire up DrawingStory state
Get viewer props from the mocked store.

* Wire up TranscribedLinesStory state
Add loadingState, stores and subjects props to TranscribedLinesStory. Pass viewer props in from store state.