Freehand Line: Continue and Undo (#2517)
* Added DragHandle to path

Initial DrahHandle will be for closing the path. The last DragHandle will be used for Continue Drawing.

* Created 'appendPath'

This allows user to continue drawing.

* Removed black 'stroke'

* WIP: work for dragHandle

* DragHandle now invisible when drawing

* When continue drawing cursor is now crosshair

* Added ability to close drawing path

Replaced the initial starting point DragHandle with a svg circle element because this handle will not be draggable.
Added isCloseToStart method which will close the drawing path if the drawing point is within 10px of the starting point.

* Code clean-up

Added PropTypes comments for Storybook documentation.

* Cleaned up deleteButtonPosition code

* Start of the Undo Feature

Created button and tweaked placement.

* Undo functionality working

Yeah!

* Added PropTypes comments

for Storybook documentation. And made Undo button focusable.

* Added README file