Refactor annotatedSteps as subject.stepHistory (#2520)
* Reorganise subject models
Move all subject code into the SubjectStore directory.

* Move store.annotatedSteps to subject.stepHistory

* Update the StepHistory tests

* Add a classification listener
Wait for a new classification to start, then start a new history for the active subject.

* Fix ClassificationStore tests
Completing a classifications triggers a reset of the store, which destroys the classification that we're trying to test. Stubbing `classifications.reset` isn't ideal, but avoids this.