Add svg line element to close polyline
This is to get around Mobx rules: will not allow adding an element that already exists in state.
Will not allow: [a, b, c, d] to become [a, b, c, d, a] (a closed path)
And using the <polygon> element seems confusing when the shape auto closes after creating 3 points.