SingleTextViewer: remove implementation tests (#2923)
Remove subject viewer tests that check details of the implementation, like `componentDidMount`, `componentDidUpdate` and Enzyme wrapper state.

Instead, test the component's rendered output and callbacks after mounting with and without a subject. This makes the tests more robust and opens up the path for converting the viewer from a class to a function with hooks.