Classifier: Set ModalTutorial as complete on close (#2924)
* Set ModalTutorial as complete on close
Call `tutorial.setSeenTime()` when the modal tutorial is closed by either activating the close button, clicking outside the popup, or pressing the `Esc` key. That marks the tutorial as complete in user project preferences.

* Remove setSeenTime from SlideTutorial
Move `tutorial.setSeenTime()` up to the `onClick` handler of the tutorial tab, in the task area.