Test the Classifier (#3046)
* Test the Classifier
Add a test that renders the classifier with a mock workflow and single image subject. Check the rendered classifier for the two task area tabs, the task answers and the SVG subject image.

* Use task answers to query radio buttons

* Test locale changes

* Add tests for inactive workflows
Mock the API for a user with and without permission to view an active workflow. Load an empty store and check that the subject and task only appear for users with permission to view inactive workflows.

* Test all the project roles

Co-authored-by: Delilah C. <23665803+goplayoutside3@users.noreply.github.com>

* Test the showTutorial prop

Co-authored-by: Delilah C. <23665803+goplayoutside3@users.noreply.github.com>