Classifier: Refactor workflow strings as a map (#3239)
* Refactor workflow strings as a map
`types.frozen` is assumed to be immutable, but `workflow.strings` changes when the locale changes. This refactors `workflow.strings` as an observable map, applying new snapshots of the language strings when the locale changes.

* Explicitly react to workflow version changes
Explicitly update the `workflowSteps` model when `workflow.version` changes.

* Remove workflowID prop from Classifier
Pass the workflow ID down via `workflowSnapshot`.

* Update autorun reactions
Remove `workflow.strings` from `_onWorkflowChange`, so that it only runs when `workflow.id` or `workflow.version` changes. Test `WorkflowStepStore` when the workflow version or language change.

* Rename `_onLocaleChange` as `_onWorkflowStringsChange`

Co-authored-by: Delilah C. <23665803+goplayoutside3@users.noreply.github.com>