Refactor ClassifierContainer effect hooks (#3273)
There's some code in `ClassifierContainer` which runs when the classifier mounts, but should only run when the window loads and the classifier store is created. This moves that code into the `useHydratedStore` hook, where it runs when the store is first created.

The effect of this is that reloading the page should give you a new subject, as expected, but unmounting and mounting the classifier by navigating the NextJS app, eg. by changing locale, should preserve your current subject and classification, even when session storage is enabled for your project.

Co-authored-by: Delilah C <23665803+goplayoutside3@users.noreply.github.com>