Generalise Classifier workflow tests (#3294)
* Generalise Classifier workflow tests
Add `Classifier.workflows.spec.js`, which defines an array of mock workflows, then loops over that array, testing the Classifier with each workflow in turn.

This tests a branching question workflow and the PH-TESS workflow at the moment. Hopefully, it can be generalised to other types of workflow eg. transcription or survey.

Co-authored-by: Delilah C. <23665803+goplayoutside3@users.noreply.github.com>