Limit 'alice' reducer to transcription workflows (#4269)
* Limit 'alice' reducer to transcription workflows
Add `workflow.caesarReducer`, which is 'alice' for transcription workflows but empty otherwise. Use that reducer when deciding whether to request Caesar reductions for a subject.

This leaves room for other reducers to be added to workflows in future.

* Generalise subject Caesar reductions
Refactor `subject.transcriptionReductions` as `subject.caesarReductions`, which can store reductions for any reducer key, if a corresponding model exists.