Refactor Panoptes authentication in all packages  (#4803)
* Get the user from the Panoptes JWT
Refactor `usePanoptesUser` to get the current user from the stored Panoptes JWT. Fall back to `auth.checkCurrent()`, which is slower, if we don't have a stored token.

* Update session auth in the classifier

* Update session auth in the content pages app

* Classifier: add user loading state
Check Panoptes user state in `ClassifierContainer`. Defer rendering `Classifier` while that check is taking place.

* Move decodeJWT to the API client

* Maintain UPP when user refreshes in the background
Setting the UPP loading state to 'loading' will unmount the classifier, so avoid that after the page has loaded and the classifier is mounted.

* Remove duplicate populateQueue on user change

* move project preferences into ClassifierContainer

* Move project roles into ClassifierContainer

* Move allowed workflows up to ClassifierContainer

* Refactor user loading state
Wait for all user session data to load before rendering the classifier. Refactor the user hooks to use the latest SWR API.

* Test admin mode in the classifier container

* Use an SWR-like API for all Panoptes user hooks

---------

Co-authored-by: Delilah C. <23665803+goplayoutside3@users.noreply.github.com>