build: ignore post-install scripts
The GitHub CI workflows and Docker builds already run `yarn install` with the `--ignore-scripts` flag. This PR makes the same change to the bootstrap scripts for local development.

I've also tidied up `bin/bootstrap.sh` so that it only runs the install once, then builds the various packages in order.